implement component for attribute field to use in row and show
