+++ frontend/app/styles/encryptable-show.scss
@@ -111,3 +111,7 @@
   display: grid;
   grid-template-columns: 6fr 1fr;
 }
+
+.encryptable-attributes > .hidden-attribute-copy-button {
+  width: 40%;
+}
