+++ frontend/app/templates/components/encryptable/attribute-field.hbs
@@ -1,0 +1,23 @@
+{{#unless this.args.row}}
+  <div id="encryptable-{{this.args.attribute}}-label">
+    {{#if this.args.label}}
+      {{this.args.label}}
+    {{else}}
+      {{t this.args.attribute}}
+    {{/if}}
+  </div>
+{{/unless}}
+<div class="hidden-attribute-copy-button">
+  {{#if this.isAttributeVisible}}
+    <div class="result-hidden-attribute">
+      <Input id="cleartext_{{this.args.attribute}}" class="text-truncate d-inline form-control {{unless this.args.row "w-100" "mt-2"}}" @disabled="true" @value={{format-cleartext this.args.encryptable this.args.attribute}}/>
+    </div>
+  {{else}}
+    <a id="show-{{this.args.attribute}}" role="button" class="text-truncate correct-pixel-error-omg show-hidden-attribute-button show-text {{unless this.args.row "w-100" "mt-2"}}" {{on "click" (action "showValue")}}>{{t (concat "encryptable/credentials.show.show_" this.args.attribute)}}</a>
+  {{/if}}
+  <CopyButton class="btn btn-light copy-btn {{if this.args.row "mt-2"}}" @clipboardText={{format-cleartext this.args.encryptable this.args.attribute}}
+              title={{if (blank-value this.args.encryptable this.args.attribute) this.noCopyBlankTooltip ""}}
+              @success={{fn this.onCopied this.args.attribute}} disabled={{blank-value this.args.encryptable this.args.attribute}} >
+    <img class="icon-clippy img d-inline" src="/assets/images/clipboard.svg" alt="clip">
+  </CopyButton>
+</div>
