+++ frontend/app/templates/components/encryptable/show.hbs
@@ -91,94 +91,22 @@
     {{/if}}
     <div class="encryptable-attributes">
       {{#if @encryptable.cleartextUsername}}
-        <div class="result-username">
-          <div id="encryptable-username-label">{{t "username"}}</div>
-          <Input id="cleartext_username" class="d-inline form-control" @disabled="true" @value={{@encryptable.cleartextUsername}}/>
-          <CopyButton class="btn btn-light copy-btn" @clipboardText={{@encryptable.cleartextUsername}}
-            title={{if @encryptable.isPasswordBlank this.noCopyBlankTooltip ""}}
-            @success={{fn this.onCopied "username"}} disabled={{@encryptable.isUsernameBlank}} >
-            <img class="icon-clippy img d-inline" src="/assets/images/clipboard.svg" alt="clip">
-          </CopyButton>
-        </div>
+        <Encryptable::AttributeField @row={{false}} @attribute="username" @encryptable={{@encryptable}} @visibleByDefault={{true}}/>
       {{/if}}
       {{#if @encryptable.cleartextPassword}}
-        <div id="encryptable-password-label">{{t "password"}}</div>
-        <div class="hidden-attribute-copy-button">
-          {{#if this.isPasswordVisible}}
-            <div class="result-hidden-attribute">
-              <Input id="cleartext_password" class="d-inline form-control" @disabled="true" @value={{@encryptable.cleartextPassword}}/>
-            </div>
-          {{else}}
-            <a id="show-password" role="button" class="correct-pixel-error-omg show-hidden-attribute-button show-text" {{on "click" (action "showValue" "password")}}>{{t "encryptable/credentials.show.show_password"}}</a>
-          {{/if}}
-          <CopyButton class="btn btn-light copy-btn" @clipboardText={{@encryptable.cleartextPassword}}
-                      title={{if @encryptable.isPasswordBlank this.noCopyBlankTooltip ""}}
-                      @success={{fn this.onCopied "password"}} disabled={{@encryptable.isPasswordBlank}} >
-            <img class="icon-clippy img d-inline" src="/assets/images/clipboard.svg" alt="clip">
-          </CopyButton>
-        </div>
+        <Encryptable::AttributeField @row={{false}} @attribute="password" @encryptable={{@encryptable}} @visibleByDefault={{false}}/>
       {{/if}}
       {{#if @encryptable.cleartextEmail}}
-        <div class="result-email">
-          <div id="encryptable-email-label">{{t "email"}}</div>
-          <Input id="cleartext_email" class="d-inline form-control" @disabled="true" @value={{@encryptable.cleartextEmail}}/>
-          <CopyButton class="btn btn-light copy-btn" @clipboardText={{@encryptable.cleartextEmail}}
-            title={{if @encryptable.isPasswordBlank this.noCopyBlankTooltip ""}}
-            @success={{fn this.onCopied "email"}} disabled={{@encryptable.isEmailBlank}} >
-            <img class="icon-clippy img d-inline" src="/assets/images/clipboard.svg" alt="clip">
-          </CopyButton>
-        </div>
+        <Encryptable::AttributeField @row={{false}} @attribute="email" @encryptable={{@encryptable}} @visibleByDefault={{true}}/>
       {{/if}}
       {{#if @encryptable.cleartextPin}}
-        <div id="encryptable-pin-label">{{t "pin"}}</div>
-        <div class="hidden-attribute-copy-button">
-          {{#if this.isPinVisible}}
-            <div class="result-hidden-attribute">
-              <Input id="cleartext_pin" class="d-inline form-control" @disabled="true" @value={{@encryptable.cleartextPin}}/>
-            </div>
-          {{else}}
-            <a id="show-pin" role="button" class="just-testing correct-pixel-error-omg show-hidden-attribute-button show-text" {{on "click" (action "showValue" "pin")}}>{{t "encryptable/credentials.show.show_pin"}}</a>
-          {{/if}}
-          <CopyButton class="btn btn-light copy-btn" @clipboardText={{@encryptable.cleartextPin}}
-            title={{if @encryptable.isPinBlank this.noCopyBlankTooltip ""}}
-            @success={{fn this.onCopied "pin"}} disabled={{@encryptable.isPinBlank}} >
-            <img class="icon-clippy img d-inline" src="/assets/images/clipboard.svg" alt="clip">
-          </CopyButton>
-        </div>
+        <Encryptable::AttributeField @row={{false}} @attribute="pin" @encryptable={{@encryptable}} @visibleByDefault={{false}}/>
       {{/if}}
       {{#if @encryptable.cleartextToken}}
-        <div id="encryptable-token-label">{{t "token"}}</div>
-        <div class="hidden-attribute-copy-button">
-          {{#if this.isTokenVisible}}
-            <div class="result-hidden-attribute">
-              <Input id="cleartext_token" class="d-inline form-control" @disabled="true" @value={{@encryptable.cleartextToken}}/>
-            </div>
-          {{else}}
-            <a id="show-token" role="button" class="correct-pixel-error-omg show-hidden-attribute-button show-text" {{on "click" (action "showValue" "token")}}>{{t "encryptable/credentials.show.show_token"}}</a>
-          {{/if}}
-          <CopyButton class="btn btn-light copy-btn" @clipboardText={{@encryptable.cleartextToken}}
-            title={{if @encryptable.isTokenBlank this.noCopyBlankTooltip ""}}
-            @success={{fn this.onCopied "token"}} disabled={{@encryptable.isTokenBlank}} >
-            <img class="icon-clippy img d-inline" src="/assets/images/clipboard.svg" alt="clip">
-          </CopyButton>
-        </div>
+        <Encryptable::AttributeField @row={{false}} @attribute="token" @encryptable={{@encryptable}} @visibleByDefault={{false}}/>
       {{/if}}
       {{#if @encryptable.cleartextCustomAttr}}
-        <div id="encryptable-custom-attr-label">{{@encryptable.cleartextCustomAttr.label}}</div>
-        <div class="hidden-attribute-copy-button">
-          {{#if this.isCustomAttrVisible}}
-            <div class="result-hidden-attribute">
-              <Input id="cleartext_custom-attr" class="d-inline form-control" @disabled="true" @value={{@encryptable.cleartextCustomAttr.value}}/>
-            </div>
-          {{else}}
-            <a id="show-custom-attr" role="button" class="correct-pixel-error-omg show-hidden-attribute-button show-text" {{on "click" (action "showValue" "customAttr")}}>{{t "encryptable/credentials.show.show_custom_attr"}}</a>
-          {{/if}}
-          <CopyButton class="btn btn-light copy-btn" @clipboardText={{@encryptable.cleartextCustomAttr.value}}
-            title={{if @encryptable.isCustomAttrBlank this.noCopyBlankTooltip ""}}
-            @success={{fn this.onCopied "customAttr"}} disabled={{@encryptable.isCustomAttrBlank}} >
-            <img class="icon-clippy img d-inline" src="/assets/images/clipboard.svg" alt="clip">
-          </CopyButton>
-        </div>
+        <Encryptable::AttributeField @row={{false}} @attribute="customAttr" @label={{@encryptable.cleartextCustomAttrLabel}} @encryptable={{@encryptable}} @visibleByDefault={{false}}/>
       {{/if}}
     </div>
     <div class="d-flex flex-row justify-content-between mb-2">
