+++ frontend/translations/fr.yml
@@ -99,13 +99,13 @@ fr:
       show_pin: Afficher Pin
       show_token: Afficher le jeton
       show_email: Afficher l'e-mail
-      show_custom_attr: Afficher l'attribut personnalisé
+      show_customAttr: Afficher l'attribut personnalisé
       copy_password: Copier le mot de passe
       copy_username: Copier le nom d'utilisateur
       copy_pin: Copier la broche
       copy_token: Copier le jeton
       copy_email: Copier l'e-mail
-      copy_custom_attr: Copier un attribut personnalisé
+      copy_customAttr: Copier un attribut personnalisé
       copy: Copier
       add_attachment: Ajouter un attachement
       blank: Vide
